backend/pyproject.toml
[project]
name = "ex1"
version = "0.3.0"
description = "Football Player Service — CRUD API with validation, tests, auth, and microservices."
requires-python = ">=3.13"
dependencies = [
    "fastapi~=0.123.0",
    "httpx~=0.28.1",
    "pydantic~=2.12.5",
    "pydantic-settings~=2.12.0",
    "slowapi~=0.1.9",
    "ruff~=0.14.8",
    "uvicorn~=0.38.0",
    "sqlmodel~=0.0.18",
    "alembic~=1.14.0",
    "psycopg2-binary~=2.9.9",
    "python-jose[cryptography]~=3.3.0",
    "passlib[bcrypt]~=1.7.4",
    "python-multipart~=0.0.18",
    "redis~=5.2.1",
    "rq~=2.1.0",
    "email-validator~=2.2.0",
]

[dependency-groups]
dev = [
    "pytest~=8.3.0",
    "pytest-asyncio~=0.24.0",
    "pytest-anyio~=0.0.0",
    "pre-commit~=3.8.0",
    "schemathesis~=3.39.0",
]
